Inclusão do Double

+++ Java-lista01-inicio/src/lista01/Lista01Exer04.java
@@ -7,7 +7,7 @@ public class Lista01Exer04 {
 	public static void main(String[] args) {
 		Scanner ler = new Scanner(System.in);
 		int a, b, c;
-		float r, s, d;
+		double r, s, d;
 		
 		System.out.print("Digite um valor inteiro para A: ");
 		a = ler.nextInt();
@@ -18,8 +18,8 @@ public class Lista01Exer04 {
 		System.out.print("Digite um valor inteiro para C: ");
 		c = ler.nextInt();
 		
-		r = (float) Math.pow((a+b), 2);
-		s = (float) Math.pow((b+c), 2);
+		r = Math.pow((a+b), 2);
+		s = Math.pow((b+c), 2);
 		d = (r + s) / 2;
 		
 		System.out.println("Valor de D é: " + d);
